Split discogs.py into smaller and more workable modules.
